bg_helpers::buffer multilinestring does eulerian_paths first
