+++ eulerian_paths.hpp
@@ -255,5 +255,11 @@ template <typename point_t, typename linestring_t, typename multi_linestring_t, 
       paths, allow_reversals).get();
 }
 
+// Returns a minimal number of toolpaths that include all the milling in the
+// oroginal toolpaths.  Each path is traversed once.  First paths are
+// directional, second are bidi.  In the pair, the first is directional and the
+// second is bidi.
+multi_linestring_type_fp make_eulerian_paths(const std::vector<std::pair<linestring_type_fp, bool>>& toolpaths);
+
 }; // namespace eulerian_paths
 #endif //EULERIAN_PATHS_H
