+++ surface_vectorial.cpp
@@ -56,7 +56,6 @@ using Glib::build_filename;
 #include "tsp_solver.hpp"
 #include "surface_vectorial.hpp"
 #include "eulerian_paths.hpp"
-#include "segmentize.hpp"
 #include "bg_helpers.hpp"
 #include "units.hpp"
 #include "path_finding.hpp"
@@ -68,9 +67,6 @@ using std::next;
 using std::dynamic_pointer_cast;
 
 unsigned int Surface_vectorial::debug_image_index = 0;
-
-// For use when we have to convert from float to long and back.
-const double SCALE = 1000000.0;
 
 Surface_vectorial::Surface_vectorial(unsigned int points_per_circle,
                                      ivalue_t min_x, ivalue_t max_x,
@@ -233,70 +229,14 @@ void Surface_vectorial::write_svgs(const string& tool_suffix, coordinate_type_fp
   }
 }
 
-// Returns a minimal number of toolpaths that include all the milling in the
-// oroginal toolpaths.  Each path is traversed once.  First paths are
-// directional, second are bidi.  In the pair, the first is directional and the
-// second is bidi.
-multi_linestring_type_fp make_eulerian_paths(const vector<pair<linestring_type_fp, bool>>& toolpaths) {
-  // Merge points that are very close to each other because it makes
-  // us more likely to find intersections that was can use.
-  auto merged_toolpaths = toolpaths;
-  merge_near_points(merged_toolpaths, 0.00001);
-
-  // First we need to split all paths so that they don't cross.  We need to
-  // scale them up because the input is not floating point.
-  vector<segment_type_p> all_segments;
-  vector<bool> allow_reversals;
-  for (const auto& toolpath_and_allow_reversal : merged_toolpaths) {
-    const auto& toolpath = toolpath_and_allow_reversal.first;
-    for (size_t i = 1; i < toolpath.size(); i++) {
-      all_segments.push_back(
-          segment_type_p(
-              point_type_p(toolpath[i-1].x() * SCALE, toolpath[i-1].y() * SCALE),
-              point_type_p(toolpath[i  ].x() * SCALE, toolpath[i  ].y() * SCALE)));
-      allow_reversals.push_back(toolpath_and_allow_reversal.second);
-    }
-  }
-  vector<std::pair<segment_type_p, bool>> split_segments = segmentize::segmentize(all_segments, allow_reversals);
-
-  // Make a minimal number of paths from those segments.
-  struct PointLessThan {
-    bool operator()(const point_type_fp& a, const point_type_fp& b) const {
-      return std::tie(a.x(), a.y()) < std::tie(b.x(), b.y());
-    }
-  };
-  // Only allow reversing the direction of travel if mill_feed_direction is
-  // ANY.  We need to scale them back down.
-  multi_linestring_type_fp segments_as_linestrings;
-  segments_as_linestrings.reserve(split_segments.size());
-  allow_reversals.clear();
-  allow_reversals.reserve(split_segments.size());
-  for (const auto& segment_and_allow_reversal : split_segments) {
-    // Make a little 1-edge linestrings.
-    linestring_type_fp ls;
-    const auto& segment = segment_and_allow_reversal.first;
-    const auto& allow_reversal = segment_and_allow_reversal.second;
-    ls.push_back(point_type_fp(segment.low().x() / SCALE, segment.low().y() / SCALE));
-    ls.push_back(point_type_fp(segment.high().x() / SCALE, segment.high().y() / SCALE));
-    segments_as_linestrings.push_back(ls);
-    allow_reversals.push_back(allow_reversal);
-  }
-
-  return eulerian_paths::get_eulerian_paths<
-      point_type_fp,
-      linestring_type_fp,
-      multi_linestring_type_fp,
-      PointLessThan>(segments_as_linestrings, allow_reversals);
-}
-
 // Make eulerian paths if needed.  Sort the paths order to make it faster.
-// Simplify paths by remving points that don't affect the path or affect it very
-// little.
+// Simplify paths by removing points that don't affect the path or affect it
+// very little.
 multi_linestring_type_fp Surface_vectorial::post_process_toolpath(
     const std::shared_ptr<RoutingMill>& mill, const vector<pair<linestring_type_fp, bool>>& toolpath) const {
   multi_linestring_type_fp combined_toolpath;
   if (mill->eulerian_paths) {
-    combined_toolpath = make_eulerian_paths(toolpath);
+    combined_toolpath = eulerian_paths::make_eulerian_paths(toolpath);
   } else {
     combined_toolpath.reserve(toolpath.size());
     for (const auto& ls_and_allow_reversal : toolpath) {
@@ -536,7 +476,7 @@ void attach_ring(const ring_type_fp& ring,
   for (const auto& ring_path : ring_paths) {
     ring_paths_with_direction.push_back(make_pair(ring_path, dir == MillFeedDirection::ANY));
   }
-  ring_paths = make_eulerian_paths(ring_paths_with_direction); // Rejoin those paths as possible.
+  ring_paths = eulerian_paths::make_eulerian_paths(ring_paths_with_direction); // Rejoin those paths as possible.
   for (const auto& ring_path : ring_paths) { // Maybe more than one if the masking cut one into parts.
     attach_ls(ring_path, toolpaths, dir, already_milled_shrunk, path_finder);
   }
