+++ testing/gerbv_example/sharp_corner_2/expected/back.ngc
@@ -163,7 +163,7 @@ G04 P0 ( dwell for no time -- G64 should not smooth over this point )
 G01 F360.00000
 G01 X-6.95682 Y-4.18327
 G01 X-6.95960 Y-4.18327
-G01 X-6.95764 Y-4.18130
+G01 X-6.95763 Y-4.18130
 G01 X-6.95828 Y-4.17916
 G01 X-6.96436 Y-4.18523
 G01 X-6.95577 Y-4.18523
